move functions to gpt_ptune_dataset
