+++ nemo/collections/nlp/data/language_modeling/megatron/t5_dataset.py
@@ -297,45 +297,3 @@ class T5Dataset(MegatronDataset):
         loss_mask = np.array(loss_mask, dtype=np.int64)
 
         return tokens_enc, tokens_dec_in, labels, enc_mask, dec_mask, loss_mask
-
-
-def make_attention_mask(source_block, target_block, pad_id):
-    """
-    Returns a 2-dimensional (2-D) attention mask
-    :param source_block: 1-D array
-    :param target_block: 1-D array
-    """
-    mask = (target_block[None, :] != pad_id) * (source_block[:, None] != pad_id)
-    mask = mask.astype(np.int64)
-    # (source_length, target_length)
-    return mask
-
-
-def make_attention_mask_3d(source_block, target_block, pad_id):
-    """
-    Returns a 3-dimensional (3-D) attention mask
-    :param source_block: 1-D array
-    :param target_block: 1-D array
-    """
-    mask = (target_block[:, None, :] != pad_id) * (source_block[:, :, None] != pad_id)
-    # (batch, source_length, target_length)
-    # mask = mask.astype(np.int64)
-    return mask
-
-
-def make_history_mask(block):
-    length = block.shape[0]
-    arange = np.arange(length)
-    history_mask = arange[None,] <= arange[:, None]
-    history_mask = history_mask.astype(np.int64)
-    return history_mask
-
-
-def make_history_mask_3d(block):
-    batch, length = block.shape
-    arange = torch.arange(length, device=block.device)
-    history_mask = (arange[None,] <= arange[:, None])[
-        None,
-    ]
-    history_mask = history_mask.expand(batch, length, length)
-    return history_mask
